Cambio de decimal a float...a rezar a lo que sea

+++ backend/database/migrations/2021_04_19_202730_alter_table_datos_mina_agregar_n_tiros_m2.php
@@ -14,7 +14,7 @@ class AlterTableDatosMinaAgregarNTirosM2 extends Migration
     public function up()
     {
         Schema::table('datos_minas', function (Blueprint $table) {
-            $table->decimal('tiros_por_m2',15,2)->default(0.00);
+            $table->float('tiros_por_m2',15,2)->default(0.00);
         });
     }
